requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.22.0
schedule==0.6.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
Werkzeug==0.15.5
